Put the NavigationBar in a new file
